Refactoring - generate.ts should be thinner
