+++ .lintstagedrc
@@ -1,4 +1,4 @@
 {
-  "*.{js,jsx,ts,tsx}": ["yarn lint:fix", "git add"],
+  "*.{js,ts}": ["npm run lint:fix", "git add"],
   "{*.{json,md}}": ["prettier --write", "git add"]
 }
