+++ packages/zubora-cli/package.json
@@ -19,7 +19,7 @@
   },
   "scripts": {
     "format": "prettier --write **/*.{js,jsx,ts,tsx,json}",
-    "lint": "eslint --ignore-path ../../.gitignore . --ext ts --ext tsx --ext js --ext jsx",
+    "lint": "eslint --ignore-path ../../.gitignore . --ext ts --ext js",
     "lint:fix": "npm run lint --fix",
     "clean-build": "rm -rf ./build",
     "compile": "tsc -p tsconfig.build.json",
